调通项目

+++ src/main/java/com/vens/mapper/UserMapper.java
@@ -5,7 +5,7 @@ import com.vens.pojos.User;
 
 public interface UserMapper {
     @DataSource("master")
-    public long add(User user);
+    public int add(User user);
     @DataSource("slave")
     public User getByPhone(String phone);
